% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeeGUTS-package.R
\docType{data}
\name{betacyfluthrinChronic}
\alias{betacyfluthrinChronic}
\title{Survival datasets for \emph{Honey bees} exposed to
constant concentration of Betacyfluthrin for 10 days.}
\format{
A list of class \code{beeSurvData} constructed by \code{dataGUTS} containing:
\describe{
\item{\code{nDatasets}}{An integer representing the number of datasets used.}
\item{\code{survData}}{A data frame containing the survival information over time
for five treatments and a control in a wide format.}
\item{\code{survData_long}}{A data frame containing the survival information over time
for five treatments and a control in a long format.}
\item{\code{concData}}{A data frame containing the concentration information over time
for five treatments and a control in a wide format.}
\item{\code{concData_long}}{A data frame containing the concentration information over time
for five treatments and a control in a long format.}
\item{\code{unitData}}{A character string containing the units of the concentration data.}
\item{\code{typeData}}{A character string containing the type of data (here Chronic_Oral).}
\item{\code{beeSpecies}}{A character string containing the species of bee of interest (here Honey_Bee).}
\item{\code{concModel}}{A data frame containing the concentration information recalculated
for the species of bee and test type of interest in a wide format.}
\item{\code{concModel_long}}{A data frame containing the concentration information recalculated
for the species of bee and test type of interest in a long format.}
\item{\code{messages}}{A data frame containing the warning messages returned by the function.}
}
}
\usage{
data(betacyfluthrinChronic)
}
\description{
Survival datasets for \emph{Honey bees} exposed to
constant concentration of Betacyfluthrin for 10 days.
}
\references{
Bayer data.
}
\keyword{dataset}
