% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{concAC}
\alias{concAC}
\title{Title}
\usage{
concAC(cExt, expTime, k_ca = 0.4)
}
\arguments{
\item{cExt}{The concentration applied}

\item{expTime}{The duration of the experiment in days}

\item{k_ca}{Contact availability rate (d-1), default is 0.4}
}
\value{

}
\description{
Title
}
\examples{
conc <- concAC(cbind(3.1, 4, 6, 8), 4)
}
