% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{concAO}
\alias{concAO}
\title{Title}
\usage{
concAO(cExt, cTime = 0.25, expTime, k_sr = 0.625)
}
\arguments{
\item{cExt}{A dataframe of concentrations at time 0 concentration applied}

\item{cTime}{The duration of exposure in days, default is 0.25 d}

\item{expTime}{The duration of the experiment in days}

\item{k_sr}{Stomach release rate (d-1), default is 0.625}
}
\value{

}
\description{
Title
}
\examples{
conc <- concAO(cExt = cbind(3.5, 6, 8, 10), cTime = 0.25, expTime = 4)
}
