% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeeGUTS.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Plotting of correlation plots for \code{beeSurvFit} objects}
\usage{
correlation_plot(object)
}
\arguments{
\item{object}{An object of class \code{beeSurvFit} to be plotted}
}
\value{
A graphic with the plot of the correlations between parameters or a
a list of graphics in case multiple datasets are present. If a list is
returned, the first graphic is the correlation between hb values and the
second is the correlation between all other parameters
}
\description{
It plots the correlations between the parameters of
the GUTS IT or GUTS SD. The plots do not include warmup iterations
}
\examples{
data(fitBetacyfluthrin_Chronic)
correlation_plot(fitBetacyfluthrin_Chronic)
}
