% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{dataGUTS}
\alias{dataGUTS}
\title{Read and format the data for the BeeGUTS model}
\usage{
dataGUTS(
  file_location = NULL,
  test_type = NULL,
  bee_species = "Honey_Bee",
  ...
)
}
\arguments{
\item{file_location}{Location of a text file containing two datasets, one for the survival data,
and one for the concentration data. Both datasets must contain the same number of column in the same order.
The following columns must be included:
\itemize{
\item \code{Survival time [d]}: a vector of time in days
\item \code{Control} A vector of number of survivors for the control
\item \code{T1} - \code{Tn} A vector of number of survivors for the treatments
T1 to Tn, one column per treatment.
}
\itemize{
\item \code{Concentration time [d]}: a vector of time in days.
\item \code{Control} A vector of concentrations for the control
\item \code{T1} - \code{Tn} A vector of concentration for the treatments
T1 to Tn, one column per treatment.
}
For the \code{Acute_Oral} and \code{Acute_Contact}, only the initial
exposure concentration at time 0 is required.}

\item{test_type}{the test type amongst "Acute_Oral", "Acute_Contact", and "Chronic_Oral"}

\item{bee_species}{the bee type. At the moment only "Honey_Bee" is supported}

\item{...}{Optional arguments to be passed to the concentration reconstruction (e.g.
\code{k_sr =} for the stomach release rate (d-1), default is 0.625,
\code{k_ca =} contact availability rate (d-1), default is 0.4), or
\code{cTime =} the duration of exposure in days for the acute oral tests, default is 0.25 d}
}
\value{
DESCRIPTION TO COMPLETE
}
\description{
DESCRIPTION TO COMPLETE
}
\examples{
\dontrun{
file_location <- system.file("extdata", "betacyfluthrin_chronic_ug.txt", package = "BeeGUTS")
lsData <- dataGUTS(file_location = file_location, test_type = 'Chronic_Oral')
}
}
