% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeeGUTS-package.R
\docType{data}
\name{fitBetacyfluthrin_Chronic}
\alias{fitBetacyfluthrin_Chronic}
\title{Model calibration results datasets for \emph{Honey bees} exposed to
constant concentration of Betacyfluthrin for 10 days.}
\format{
A list of class \code{beeSurvFit} constructed by \code{fitBeeGUTS} containing:
\describe{
\item{\code{stanFit}}{A 'stanfit'  object containing the results of the calibration.}
\item{\code{data}}{A 'beeSurvData' objects with the user data used for the calibration.}
\item{\code{dataFit}}{A list containing the priors and data formatted for the calibration algorithm.}
\item{\code{setupMCMC}}{A list containing the setup used for the MCMC.}
\item{\code{modelType}}{A character string containing the type of GUTS model used (here 'SD').}
\item{\code{distribution}}{A character string containing the distribution used (IT only, here 'NA').}
\item{\code{messages}}{A character string containing the error messages if Rhat >1.1 (here 'NA').}
}
}
\usage{
data(fitBetacyfluthrin_Chronic)
}
\description{
Model calibration results datasets for \emph{Honey bees} exposed to
constant concentration of Betacyfluthrin for 10 days.
}
\references{
Bayer data.
}
\keyword{dataset}
