% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCx.R
\name{LCx}
\alias{LCx}
\title{Predict the Lethal Concentration method at which \eqn{x}\\% of organisms die for any
specified time-point for a \code{beeSurvFit} object}
\usage{
LCx(object, ...)
}
\arguments{
\item{object}{An object used to select a method}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
Predict median and 95\\% credible interval of the x\\% Lethal Concentration.
}
\details{
When class of \code{object} is \code{beeSurvFit}, see \link[=LCx.beeSurvFit]{LCx.beeSurvFit}.
}
