% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeeGUTS.R
\name{plot.ppc}
\alias{plot.ppc}
\title{Plotting method for \code{ppc} objects}
\usage{
\method{plot}{ppc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ppc}.}

\item{\dots}{Further arguments to be passed to generic methods.}

\item{data_type}{A string designating the type of data to be plotted: \code{length},
\code{reproduction} or \code{exposure}}
}
\value{
an object of class \code{ggplot}.
}
\description{
Plotting method for \code{ppc} objects
}
