% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictBeeGUTS.R
\name{predict.beeSurvFit}
\alias{predict.beeSurvFit}
\title{Predict method for \code{beeSurvFit} objects}
\usage{
\method{predict}{beeSurvFit}(object, dataPredict, ...)
}
\arguments{
\item{object}{An object of class \code{beeSurvFit}}

\item{dataPredict}{Data to predict in the format XXXXXXXXXXXXXXXXXXXXXXXXXXXXX}

\item{...}{Additional arguments to be parsed to the  \code{predict.survFit} method from \code{morse} (e.g.
\code{mcmc_size = 1000} is to be used to reduce the number of mcmc samples in order to speed up
the computation. \code{mcmc_size} is the number of selected iterations for one chain. Default
is 1000. If all MCMC is wanted, set argument to \code{NULL}.,
\code{hb_value  = FALSE} the background mortality \code{hb} is taken into account from the posterior.
If \code{FALSE}, parameter \code{hb} is set to a fixed value. The default is \code{FALSE}.
\code{hb_valueFORCED  = 0} hb_valueFORCED If \code{hb_value} is \code{FALSE}, it fix \code{hb}. The default is \code{0}}
}
\value{

}
\description{
This is the generic \code{predict} S3 method for the \code{beeSurvFit}
class. It predict the survival over time for the concentration profiles entered by the user.
No concentration reconstructions are performed here.
}
\examples{
\dontrun{
dataPredict <- data.frame(time = c(1:10, 1:10, 1:10), conc = c(rep(5, 10), rep(10, 10), rep(15, 10)), replicate = c(rep("rep1", 10), rep("rep2", 10), rep("rep3", 10)))
data(fitBetacyfluthrin_Chronic)
prediction <- predict(fitBetacyfluthrin_Chronic, dataPredict)
}
}
