% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeeGUTS.R
\name{priorposterior_plot}
\alias{priorposterior_plot}
\title{Plotting of prior and posterior plots for \code{beeSurvFit} objects}
\usage{
priorposterior_plot(object)
}
\arguments{
\item{object}{An object of class \code{beeSurvFit} to be plotted}
}
\value{
A list of graphic objects with the plot of the comparison between
prior and posterior distribution of the parameters. The first entry of the
list is dedicated to the hb value (multiple plots in case of multiple
datasets), the second element of the list is dedicated to the other model
parameters
}
\description{
It plots the comparison between the prior and the posterior
distributions of the parameters of the GUTS IT or GUTS SD models.
The plots do not include warmup iterations
}
\examples{
data(fitBetacyfluthrin_Chronic)
priorposterior_plot(fitBetacyfluthrin_Chronic)
}
