% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortTimeEffects.R
\name{ShortTimeEffects.beeSurvFit}
\alias{ShortTimeEffects.beeSurvFit}
\title{Calculate the presence of Time Reinforced Toxicity for the compound from the
calibrated model \code{beeSurvFit} object.}
\usage{
\method{ShortTimeEffects}{beeSurvFit}(object, concRange = NULL, fullcalculation = FALSE)
}
\arguments{
\item{object}{An object of class \code{beeSurvFit}}

\item{concRange}{Argument of LCx, range of concentrations to find LDD50}

\item{fullcalculation}{Compute the LDD50 from day 1 to day 10 of the Chronic test.
This can increase the computation time}
}
\value{
A object of class \code{ggplot} containing the graph of the comparison
between LDD50 at day 2 and day 10 and the data.frame with the plotted values.
}
\description{
Calculate the presence of Time Reinforced Toxicity for the compound from the
calibrated model \code{beeSurvFit} object.
}
\examples{
\donttest{
data(fitBetacyfluthrin_Chronic)
ShortTimeEffects(fitBetacyfluthrin_Chronic)
}
}
