% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBeeGUTS.R
\name{summary.LCx}
\alias{summary.LCx}
\title{Summary of \code{LCx} objects}
\usage{
\method{summary}{LCx}(object, ...)
}
\arguments{
\item{object}{An object of class \code{LCx}}

\item{...}{Additional arguments to be parsed to the generic \code{summary} method (not used)}
}
\value{
A summary of the \code{LCx} object
}
\description{
This is the generic \code{summary} S3 method for the \code{LCx} class.
It shows the median and 95\% credible interval of the calculated LCx.
}
\examples{
\donttest{
data(fitBetacyfluthrin_Chronic)
out <- LCx(fitBetacyfluthrin_Chronic)
summary(out)
}
}
