% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{concAC}
\alias{concAC}
\title{Recalculate the concentrations for the acute contact tests for bees}
\usage{
concAC(cExt, expTime, k_ca = 0.4, ...)
}
\arguments{
\item{cExt}{The concentration applied}

\item{expTime}{The duration of the experiment in days}

\item{k_ca}{Contact availability rate (d-1), default is 0.4}

\item{...}{Not used}
}
\value{
A data frame containing a column with the time points and a column with the
recalculated concentrations
}
\description{
Recalculate the concentrations for the acute contact tests for bees
}
\examples{
conc <- concAC(cbind(3.1, 4, 6, 8), 4)
}
