% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBeeGUTS.R
\name{fitBeeGUTS}
\alias{fitBeeGUTS}
\title{Fit a GUTS model for bees survival analysis using Bayesian Inference (stan)}
\usage{
fitBeeGUTS(
  data,
  modelType = NULL,
  distribution = "loglogistic",
  priorsList = NULL,
  parallel = TRUE,
  nCores = parallel::detectCores() - 1L,
  nChains = 3,
  nIter = 2000,
  nWarmup = floor(nIter/2),
  thin = 1,
  adaptDelta = 0.95,
  odeIntegrator = "rk45",
  relTol = 1e-08,
  absTol = 1e-08,
  maxSteps = 1000,
  ...
)
}
\arguments{
\item{...}{Additional parameters to be passed to \code{sampling} from \code{stan}}
}
\value{

}
\description{
DESCRIPTION TO COMPLETE
}
\examples{
\dontrun{
data(betacyfluthrinChronic)
fit <- fitBeeGUTS(betacyfluthrinChronic, modelType = "SD", nIter = 1000)
}
}
