% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{ppc}
\alias{ppc}
\title{Generates an object to be used in posterior predictive check for \code{beeSurvFit}, \code{beeSurvPred}}
\usage{
ppc(x)
}
\arguments{
\item{x}{an object used to select a method \code{ppc}}
}
\value{
a \code{data.frame} of class \code{ppc}
}
\description{
Generates an object to be used in posterior predictive check for \code{beeSurvFit}, \code{beeSurvPred}
}
