% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBeeGUTS.R
\name{summary.beeSurvFit}
\alias{summary.beeSurvFit}
\title{Summary of \code{beeSurvFit} objects}
\usage{
\method{summary}{beeSurvFit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{beeSurvFit}}

\item{...}{Additional arguments to be parsed to the generic \code{summary} method (not used)}
}
\value{
A summary of the \code{beeSurvFit} object
}
\description{
This is the generic \code{summary} S3 method for the \code{beeSurvFit} class.
It shows the quantiles of priors and posteriors on parameters.
}
\examples{
\donttest{
data(fitBetacyfluthrin_Chronic)
summary(fitBetacyfluthrin_Chronic)
}
}
