% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeReinforcedToxicity.R
\name{TRT}
\alias{TRT}
\title{Perform the extrapolation of the LDD50 to 27 days and check the presence of
reinforced toxicity as for EFSA revised guideline (2023) - Annex G}
\usage{
TRT(object)
}
\arguments{
\item{object}{An object used to select a method}
}
\value{
A \code{ggplot} object with graph of the LDD extrapolation compared
to the Haber's law and a data.frame with the calculations
}
\description{
Copyright: C. Romoli, ibacon GmbH, 2023-2024
}
\details{
When class of \code{object} is \code{beeSurvFit}, see \link[=TRT.beeSurvFit]{TRT.beeSurvFit}.
}
